move HStoreField into nav.django.forms
